Add a vanilla html version of HAR Viewer
Copy the php files to html files, and replace the php include statements
with the contents of the new "ga.include" file.

"loader.php" has been changed to accept an "ext" URL parameter (for
extension), and this is used to determine which version of files to load.
The php version will remain the default, but passing "ext=html" as a URL
parameter to loader.php/loader.html will pick the vanilla html version.

"har.js" has a similar change, it now also looks for a "data-ext"
attribute on the 'har element', and this is used to generate either a
"preview.php" or "preview.php" URL.

The overall idea is that a vanilla HTML version of HAR Viewer can be
deployed, e.g. to GitHub Pages.